+++ model/Helper/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+	Validator: require('./Validator'),
+	throwError: require('./throwError')
+}
