+++ model/Schema/index.js
@@ -1,0 +1,7 @@
+
+module.exports = {
+	Type: require('./Type'),
+	Nested: require('./Nested'),
+	Populate: require('./Populate'),
+	Updated: require('./Updated'),
+}
