+++ package.json
@@ -18,12 +18,13 @@
 		"chai": "^4.2.0",
 		"debug": "^4.1.1",
 		"dotenv": "^8.1.0",
-		"jest": "^24.9.0",
-		"mocha": "^6.2.1",
-		"redis": "^2.8.0",
 		"eslint": "^4.19.1",
 		"eslint-plugin-array-func": "^3.1.0",
 		"eslint-plugin-promise": "^4.2.1",
-		"eslint-plugin-require-in-package": "^1.0.3"
+		"eslint-plugin-require-in-package": "^1.0.3",
+		"expect": "^26.4.2",
+		"jest": "^24.9.0",
+		"mocha": "^6.2.1",
+		"redis": "^2.8.0"
 	}
 }
